+++ src/components/PostImage/index.tsx
@@ -1,5 +1,7 @@
 import styled from "@emotion/styled";
 
+import loadingImg from '../../assets/loading.gif';
+
 export const DEFAULT_POST_URL = "https://via.placeholder.com/115x110";
 export const DEFAULT_POST_URL_MD = "https://via.placeholder.com/400x400";
 export const DEFAULT_POST_ALT = "Image placeholder";
@@ -14,6 +16,7 @@ export interface PostImageProps extends PostImageBaseProps {
   url?: string
   alt?: string
   onImageClick?: () => void
+  loading?: boolean
 }
 
 interface PostImageBaseProps {
@@ -24,14 +27,16 @@ export function PostImage({
   url,
   alt,
   onImageClick,
-  size = PostImageSize.sm
+  size = PostImageSize.sm,
+  loading = false
 }: PostImageProps) {
 
   const finalUrl = url || (size === PostImageSize.sm ? DEFAULT_POST_URL : DEFAULT_POST_URL_MD)
 
+  if (loading) return <StyledLoadingImage alt="Loading..." src={loadingImg} />
+
   return (
     <div data-testid="post-image-root" onClick={onImageClick}>
-
       <StyledImage data-testid="post-image" src={finalUrl} alt={alt} size={size} />
     </div>
   );
@@ -42,4 +47,10 @@ const StyledImage = styled.img`
   max-height: 100%;
   margin-right: 12px;
   margin-top: 12px;
+`;
+
+const StyledLoadingImage = styled.img`
+  width: 200px;
+  height: 200px;
+  margin: 24px;
 `;